Cap bufsiz variable (long) to avoid overflow in Alphabit & Rabbit
With nb > 2^(31+5), bufsiz is not correctly computed (can lead to a negative value); furthermore, with nb > 2^(31-2), util_Min() in ufile_CreateReadBin() could miscalculate the minimum value and the following util_Calloc() returns an error. To avoid all of this, bufsiz has been limited to LONG_MAX/4 (in any case, the maximum allocated buffer will be ARRAYDIM << LONG_MAX/4).