Avoid overflow of "n" variable (long) in DoWalk()
when nb is big, it is not safe to calculate nb / (L * N) and assign it to a long variable (n), since it is possible to overflow the 2^31 value limit. The following util_Min() can produce a wrong value. To avoid this, the calculation is performed in floating point, directly as an argument of the util_Min() function.